refactor(graph): name the field after what it carries
The value is the audit the server ran, not a score it asserts, so the field is
`audit` and the function behind it is `auditResult`.
