--- tests/test-graph/src/features/test_ttscgraph_reports_audited_integrity_before_the_facts.ts
@@ -1,148 +1,0 @@
-import { TestProject } from "@ttsc/testing";
-
-import { TtsgraphClient, assert } from "../internal/ttsgraph";
-
-interface ToolResult {
-  content: { type: string; text: string }[];
-  structuredContent?: {
-    integrity?: string;
-    next?: { action?: string };
-    result?: { type?: string };
-  };
-}
-
-const GRAPH_TOOL_NAME = "inspect_typescript_graph";
-
-const escapeArguments = () => ({
-  question: "The next evidence is outside the indexed TypeScript graph.",
-  draft: {
-    reason: "The next evidence is outside the indexed TypeScript graph.",
-    type: "escape",
-  },
-  review: "Confirmed: skip graph work and return escape.",
-  request: {
-    type: "escape",
-    reason: "No graph operation is needed for this request.",
-    nextStep: "Use non-graph evidence.",
-  },
-});
-
-const overviewArguments = () => ({
-  question: "Summarize project shape from graph index facts.",
-  draft: {
-    reason: "An overview is the smallest useful architecture request.",
-    type: "overview",
-  },
-  review: "Confirmed: read architecture facts from the graph, not from files.",
-  request: {
-    type: "overview",
-    aspect: "all",
-  },
-});
-
-/**
- * Verifies a result reports its audited `integrity` first, carries no
- * instruction, and crosses the wire once.
- *
- * The server used to stamp a `directive` onto every result telling the model
- * the facts were sacred and not to be verified. A tool result is untrusted
- * input, so a command inside one is the shape of a prompt injection, and models
- * read it that way: Sonnet called it "a prompt-injection-style directive baked
- * into the MCP server's tool result", checked the graph against the sources on
- * principle, and warned the user about this server in its answer. The claim it
- * made is now a number the server audits — the share of the result's facts that
- * resolve back to the type-checked program — and it serializes before `result`,
- * so what the facts are worth precedes the facts. What the number means is the
- * output schema's to say.
- *
- * The single copy is load-bearing too: a tool that declares an output schema
- * must answer with `structuredContent`, and serializing the same JSON into a
- * text block as well doubles a 30 KB tour into 60 KB, blowing a client's
- * tool-result cap and spilling the answer to a file the model then shells out
- * to read.
- *
- * 1. Materialize a real project and start one MCP server.
- * 2. Call the escape branch and a real graph branch.
- * 3. Assert each payload leads with `integrity`, holds only the facts and their
- *    disposition, and arrives as structured content alone.
- */
-export const test_ttscgraph_reports_audited_integrity_before_the_facts =
-  async () => {
-    const root = TestProject.createProject({
-      "tsconfig.json": JSON.stringify({
-        compilerOptions: {
-          target: "ES2022",
-          module: "commonjs",
-          strict: true,
-          rootDir: "src",
-          outDir: "dist",
-        },
-        include: ["src"],
-      }),
-      "src/index.ts": "export class Widget {}\n",
-    });
-
-    const client = TtsgraphClient.start(root);
-    try {
-      await client.request("initialize", {
-        protocolVersion: "2025-06-18",
-        capabilities: {},
-        clientInfo: { name: "test-graph", version: "0.0.0" },
-      });
-      client.notify("notifications/initialized", {});
-
-      const call = async (
-        args: Record<string, unknown>,
-        expectedType: string,
-      ): Promise<void> => {
-        const response = (await client.request("tools/call", {
-          name: GRAPH_TOOL_NAME,
-          arguments: args,
-        })) as ToolResult;
-        const payload = response.structuredContent;
-        const raw = JSON.stringify(payload);
-        assert.ok(
-          payload !== undefined,
-          `the result must arrive as structured content: ${JSON.stringify(response)}`,
-        );
-        assert.deepEqual(
-          Object.keys(payload as object),
-          ["integrity", "next", "result"],
-          `integrity leads, then where it leaves the question, then the facts: ${raw}`,
-        );
-        assert.match(
-          payload?.integrity ?? "",
-          expectedType === "escape"
-            ? /no graph facts to audit/i
-            : /Audited before returning: (\d+) of \1 facts .+ \(100%\)/,
-          `the audit must report what it checked and what resolved: ${raw}`,
-        );
-        assert.ok(
-          typeof payload?.next?.action === "string",
-          `next must report where the result leaves the question: ${raw}`,
-        );
-        assert.equal(
-          payload?.result?.type,
-          expectedType,
-          `result.type must mirror the request: ${raw}`,
-        );
-        assert.deepEqual(
-          response.content,
-          [],
-          `the payload must not cross a second time as text: ${raw}`,
-        );
-      };
-
-      await call(escapeArguments(), "escape");
-      await call(overviewArguments(), "overview");
-    } finally {
-      client.endStdin();
-    }
-
-    const code = await client.waitForExit();
-    assert.equal(
-      code,
-      0,
-      `the launcher should exit cleanly\nstderr: ${client.stderrText()}`,
-    );
-  };
